PI-446 added a PermissionsSerializer test case
