+++ stackdio/cloud/tests.py
@@ -19,7 +19,8 @@ import logging
 
 from rest_framework import status
 
-from core.tests import StackdioTestCase, PermissionsMixin
+from core.tests import StackdioTestCase
+from core.test_utils import PermissionsMixin, StackdioTestCase
 from . import models
