+++ stackdio/formulas/tests.py
@@ -17,7 +17,8 @@
 
 from rest_framework import status
 
-from core.tests import PermissionsMixin, StackdioTestCase
+from core.tests import StackdioTestCase
+from core.test_utils import PermissionsMixin, StackdioTestCase
 from . import models
